Display model's class and id when to_s not defined
https://trello.com/c/9OfUvC3x

Why?

Administrate relied on `#to_s` to display objects throughout the system,
but in order for it to work correctly developers needed to define #to_s
on each model that they wanted to use with Administrate.

No more.

Now, if a developer doesn't define a custom #to_s on a model,
Administrate will detect that and default to showing the object's class
and database id.

TODO:

refactor... especially HasMany and BelongsTo relationships.