Add a generator for field partials
https://trello.com/c/wadhSHmo

Problem:

If a developer wants to change how a certain type of attribute
appears in the admin dashboards, they need to overwrite specific views.

At the moment, this requires them to dig into Administrate's source code
to find the current implementation,
and then copy that into the appropriate directory in their own app.

Solution:

Add a generator to automate the partial generation process.
Generator API is as defined at
https://administrate-docs.herokuapp.com/customizing_attribute_partials