Rearrange git repo so bundler can install from git
Problem:

Because `administrate.gemspec` was not at the root of the repository,
users could not install the gem by pointing their Gemfile at Github.

Solution:

Move the gem to the root directory and moves the example app to a
directory within the `/spec` directory. This makes it possible to refer
to the gem directly from git or Github.

The example app is stored as a [git subtree], which means it can be
deployed separately to Heroku or other hosting services.

[git subtree]: https://blogs.atlassian.com/2013/05/alternatives-to-git-submodule-git-subtree/

Change notes:

- The `translate` method defined in the controller was causing i18n-tasks
to get confused about the missing keys in the locale file. Renamed it to
`translate_with_resource`.

- Fix Capybara Webkit deprecation warning
- Load schema before running specs
- Update the example apps Gemfile to point at Github, so the example app
  can be deployed to Heroku. For local development, the local copy of
  `administrate` takes precedence over the example app's Gemfile, and
  tests still run as expected.
- Add rake task to deploy example app. This uses
  more-complex-than-normal git subtree commands, so it's good to have a
  shortcut.