Improve spacing around last link in sidebar
Problem:

Links in the sidebar have no bottom padding,
so the last link in the list has its baseline
at the exact bottom of the page.

This does not look good,
and it is easy for the last link to be obscured.

Solution:

Add padding to the bottom of the last link in the sidebar.