Display helpful message if attribute_name missing
# Problem
If a field in `COLLECTION_ATTRIBUTES` has no key in `ATTRIBUTE_TYPES` either because I misspelled it or I forget to add it to `ATTRIBUTE_TYPES` I will get a undefined method new for `nil:NilClass` in `Page::Base`

# Solution
Don't access hash values directly but use `fetch`. If the attribute is not found the block raises an error with a helpful error message.

references #246