[OPTIM] Speed up selective key fetches from hashes
Pull Request #293

Problem:

The `ATTRIBUTES_TYPES` hash is being iterated over twice
to determine which values are searchable.

This is an inefficient method
for selecting searchable keys from the hash.

Solution:

Refactor the algorithm
to reduce the time needed to iterate through the hash.

Benchmarks show that the new version is a 22% speedup
from the original.

Several other methods were tried,
as outlined in [this benchmarking script][bench].
Two options gave better than 22% speedup,
but were rejected because they reduced the readability of the code.

[bench]: https://gist.github.com/graysonwright/90017f7d74f19e80b789