Display Enum fields as Strings instead of Numbers
Problem:

ActiveRecord enums are integers in the database,
but ActiveRecord translates them into the associated string value
before returning them to Rails application code.

The administrate generator interprets Enums as integers by default,
which means that any generated dashboards with enums error
unless you manually change the attribute type to Field::String.

Solution:

When generating a dashboard class for an ActiveRecord model,
inspect the class to see if it has defined enums
per the ActiveRecord 4.1+ API.
If the column in question is an Enum,
by default interpret it as a String field that is not searchable.

As a follow-up task, we can define a custom field
for Enums that supports searching and other operations.