Add search support for ActiveRecord::Enum columns
Bit of a work in progress/hack. Map all searchable fields into a hash and
allow each field class to customize the search predicate and comparison
term based on the passed search string. If no matching enum value is
found in the search term (based on == compare), omit enum field from
search string. Otherwise, try to map it to the correct integer value.