Generate SHOW_PAGE_ATTRIBUTES as an explicit array
Closes #130

Problem:

The dashboard generator specifies:

```ruby
SHOW_PAGE_ATTRIBUTES = ATTRIBUTE_TYPES.keys
```

This means that the order of the show page attributes
is dependent on the order of keys in the `ATTRIBUTE_TYPES` hash.

Ruby hashes happen to be ordered,
but we shouldn't make an assumption that hashes are ordered
because that is not a property that hashes are designed to have.

Solution:

Because we assume that users will be editing `SHOW_PAGE_ATTRIBUTES`
to customize the contents and order of show page attributes,
we'll generate an explicit array for them to modify.