Allow linkable element on row and shouldn't go to show page
The previous problem that I had when I tried to have a column which is
clickable user's thumbnail that should go to their fecebook profile.
Unfortunately, when user clicks on thumbnail it opens facebook profile
page which is correct, but it also redirects user to show page as always
when user clicks on row.

The solution, whenever the column marks with `cell-data--linkable`, then
we assume content in that column will contain linkable elements which we
don't need to have the behaviour that always go to show page when they
click on that row.