Fix UI regression from moving the header button
Problem:

When merging in #214,
the Percy.io visual regression tests weren't run
because our quota for the month had been used up.

In that PR, we re-aligned the button in the header.
This change accidentally affected content below the header.

Solution:

Change the alignment from `baseline` to `flex-start`.
This keeps the button from affecting content below the header.