Integration with Globalize
Problem
If you are using Globalize to translate a Dashboard's resource you will only have access to the default locale values.

Solution
If a model responds to the ".translates?" class method and it's true then links to the other locales present in I18n.available_locales will be shown in the index, show and edit views.