[JRO] Only list resources if they're in the manifest
If the dashboard manifest lists out the resources `administrate` is
working with, then the links in the sidebar should only link to those
resources.

As of this commit Administrate::Namespace was inferring which resources
to list based on what routes had "#{namespace}/" in the path. If you add
a route in the namespace for a resource you do NOT want to show up, it
is possible that it would break.