Remove Neat
We are not really utilizing or in need of a full blown grid system.
Given that, and that this is an open-source project that is added to
existing projects, reducing this dependency offers nice benefits.

For now, this replaces Neat's mixins with properties to re-create a near
identical output. It would be a great next step to flesh out a *small*
class-based grid *system* for Administrate.

Related: https://github.com/thoughtbot/administrate/pull/789