Add labeled and mix context searching
Searches can be labeled with their attribute or without an attribute
they will be grouped in the "special" :all label that will be run
against any searchable attribute. Labeled searches will always be
preferred to the :all search. In addition, labels can be specified more
than once to enable searching for multiple values within the attribute.
Finally, "op: and" can be used to turn the default "OR" operator to an
"AND" operator when creating the outer query conditions (multiple
occurrences of a labeled search will always use an "OR" search unless
overridden by the search definition).

Examples:
  On the Line Items page of the same application this search will find
  all line items which are either for The Game of Life product, have a
  unit price greater than or equal to $63.00, or whose order comes to
  a total price greater than or equal to $190.00.

    `product: Game of Life, unit_price: 63, total_price: 190`

  This search is similar to the above but uses "AND" for the outer
  query, and thus requires all of the search terms to be true.

    `op: and, product: Game of Life, unit_price: 63, total_price: 190`