Add search filters to dashboards
If we add:

    COLLECTION_FILTERS = {
      inactive: ->(resources) { resources.where("login_at < ?", 1.week.ago) }
    }

to a dashboard, we can query the resources of that dashboard with

    bob inactive:

to find users named "bob" who hasn't logged in the last week.

If you already had the `inactive` scope you could define the filter like so to
take advantage of existing ActiveRecord scopes (and other class methods on the
resource class).

    COLLECTION_FILTERS = {
      inactive: ->(resources) { resources.inactive }
    }

While the chosen hash-based syntax is a bit more verbose than simply exposing
already defined scopes like so:

    # app/dashboards/customer_dashboard.rb
    COLLECTION_FILTERS = [:inactive]

it allows us to define filters for use in Administrate without having to clutter
the resource classes with scopes.

It still allows us to add the simpler syntax in a backwards compatible way at
some point down the line if we feel the need. For example it could end up
looking like:

    COLLECTION_FILTERS = {
      vip: :vip, # This could call the method `vip` on resources
      inactive: ->(resources) { resources.where("login_at < ?", 1.week.ago) }
    }

* Allow search_spec to be run on its own,
* Introduce the concept of a search query
  - adding collection scopes/filters means we need to add more involved
    search query parsing; this gives us a place for that,