Don't make dashboards with an uninitialized name
When using the `rails g administrate:dashboard` generator with an
uninitialized model name, the method will raise a NameError but the
`app/dashboards/#{foo}_dashboard.rb` file will be created regardless.
This happens because the generator uses the built-in `file_name`
parameter to create the file name regardless of whether or not the rest
of the generator runs.

To solve this, we simply check whether or not `klass` is truthy before
actually generating the file because this is the same guard that we use for the
rest of the generator.