Improve german translation.
In order to be able to improve the german translation of "New post",
the string contatenation of "new" + resource had to be moved into
the i18n string instead of being performed inside the template. This
will help other locales as well I assume. To keep the current
behaviour for all other locales, I appended "%{resource}".