Allow parameters that are more permissive
When using the Clearance gem's backdoor to access views in
administrate, tests prematurely break because of the `as` parameter
used to do quick sign in. While it's probably fair that we don't want
persist that parameter through the GET request, it shouldn't matter
what parameters we pass in a GET request.

* Use `permit!` to allow anything through on a GET request