Use show page attributes when rendering  has_one
On the show page for a record, we were previously displaying the form
attributes when rendering the child of a has_one relationship.

This makes us display the show page attributes, fitting better when rendering
a record on a show page.

Furthermore, using the form attributes can lead to displaying odd attributes
such as "Password Confirmation".