Introduce `Fields::Url`
This commit introduces a `Fields::Url` field.

When used in a form, the field will be rendered as an `<input
type="url">` element.

When used on `show` and `index` pages, the field will be rendered as an
`<a href="..."></a>` element.

Additionally, ensure that the fields are searchable and truncated to 50
characters, by default.