Be able to double-click to select attribute text
Each show page displays a description list of the fields.
When the field is text, it is rendered inside an inline `<span>`.

I've found it common to want to double-click this text
to copy-paste an identifier from our admin app into other systems.

Previously, double-clicking would not select the field's value.
The bug can be seen on the show page in the demo app. For example,
try selecting the city or zip code values from:

https://administrate-prototype.herokuapp.com/admin/orders/198962

Changing the markup to be a block `<div>` fixes the bug.
Alternatively, we could change the `preserve-whitespace` CSS rule
to be `display: block;`.