Rewrite the contributing guide
The goal behind this document should be to explain to new and potential
contributors what they need to do (and follow) to get involved and it
wasn't really doing that.

It should also be a document that we can link to when closing or explaining
reasoning behind decisions. Broadly, it should aim to ensure decisions
aren't just arbitary.

* Link to the contributing guide in the docs,
* Move repository/architecture discussion to the Contributing Guide,
* Document the use of labels (this matches those setup in GitHub),
* Document how releases happen and how it fits the hosted applications.