More natural tabbing across rows in table
When the user presses tab, the browser is expected to move to the next
link. Since all cells in the table are links, the user is expected to
press many times (once per column) in order to move to the next row.

Since each row works as a link, so much tabbing should not be necessary.
This change makes it so that each row is tabbable, but not the cells.

An exception is if the data in the cell is internally represented as a
link. This is the case of email fields. In this case, this data link is
still accessible with an additional tab press for each row: the row will
be highlighted first, then each data link.