`protected` visibility is discouraged
Coming from other languages, it's tempting to use `protected` in Ruby.
However this keyword has different nuanced in Ruby and is generally
discouraged. It should only be used for methods required for comparators
and operators. None of the methods marked as `protected` here fall into
this category, so I think they should be marked `private` instead.