Resources can be decorated
Add the possibility to decorate resources using the Decorator pattern.

This could be eg. by wrapping the resource in a class which rewrites or
adds to the field values it exposes, or by including a module which
achieves a similar effect.

This started out as an attempt to resurrect
https://github.com/thoughtbot/administrate/pull/843 by @pustomytnyk, and
a lot of that code is still in this commit. However some changes of
approach have been made, and the necessary preliminary refactorings have
been pulled out into separate commits.

NOTE: we need to expose two versions of the resource: one which is
decorated (which might be any arbitrary class) and one for pagination.

We also need to make sure we pass the undecorated resource to
`display_resource`, so that the class name of the resource gets shown
and not the class name of whatever `display_resource` returns. There's a
trade-off here: this means that the default `display_resource` works,
but that a modified `display resource` won't be able to use methods from
the decorator. This could be solved by first decorating the resource in
the custom `display_resource` definition.

I guess there's a case for having the naming the other way round: the
basic resource named `resource` and the decorated one being called
`decorated_resource`. It feels like there are trade-offs with both
options.