Refactor: resources exposed via Collection page
All the other controller actions expose the `resource` via various`Page`
objects, but for the index action `resources` has so far been exposed
directly.

We want to move towards being able to modify how resource data is
presented, and this feels like it should be a responsibility of the
`Page` objects, so pulling `resources` inside
`Adminstrate::Page::Collection` moves us towards that goal. It also
feels more consistent with the design of the other `Page` objects

This changes the interface of the Collection page object, but it seems
like this class isn't part of the public API.