Refactor: HasMany field doesn't expose resources
We're now at a point where we're asking the field for resources, and
using those to build the associated collection. This feels like a clear
violation of Tell Don't Ask.

Instead we can build the resources internally based on the order and page
number.

Note that the approach in the view partials is to check `.data.any?`
rather than `.resources.any?` since the latter no longer exists at this
point. These should return the same result, since `.resources` is just
`.data` with ordering and includes applied to them.

Since `resources` is no longer called externally, we can make it
private.  In order to do so we need to adjust a number of the specs
which asserting properties of `resources` directly. We can change these
to assert based on the resources exposed by the associated collection.
This is not ideal, but doesn't feel awful either?

Also we can no longer get away with passing an empty array as a test
double for an ActiveRecord association (which is what a Field is
actually initialized with) in our specs. For simplicity, we can use
classes that already have corresponding database tables, and use `.none`
to return the empty association.

We also can't really get away with having a test double in place of an
Administrate::Order, since it's now being applied to a record, so let's
use a real one instead.