Move field requireness logic to Field::Base
Given that application helpers are available throughout the app and the
highly specific nature of some of the actions required to determine
whether or not a field is required, moving that logic to Field::Base
feels more appropriate.

This also allows us to add more complex checks to determine whether or
not a field is required without exposing that logic everywhere in the
application.