Less confusing example of authorization
Currently, the route `payments#index` exists but is disallowed to all
users. This is probably in order to provide a context to the
authorization spec, but I think it's confusing.

Now that we have an authentication mechanism, we can use this to provide
a clearer (I hope) authorization spec and a less confusing app.