Update front_matter_parser gem from 1.0.0 to 1.0.1
Update "front_matter_parser" to be compatible with Psych 4.
Psych 4 is the version included in Ruby 3.1.
See PR 11 in the "waiting-for-dev/front_matter_parser" repo.

"front_matter_parser" was explicitly locked at 1.0.0 in Administrate in PR 2023
to avoid an incompatibility with Ruby 2.5.
Then it was unlocked in PR 2027 when support for Ruby 2.5 was dropped.
However, front_matter_parser was also dependabot-ignored in PR 2021, and I think
it was never un-ignored, so dependabot hasn't suggested an upgrade.